confirmation alert met bevestiging toegevoegd

+++ src/main/java/controller/ManageUsersController.java
@@ -4,12 +4,14 @@ import database.mysql.DBAccess;
 import database.mysql.UserDAO;
 import javafx.fxml.FXML;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TextField;
 import model.User;
 import view.Main;
 
 import java.util.ArrayList;
+import java.util.Optional;
 
 public class ManageUsersController {
 
@@ -47,11 +49,17 @@ public class ManageUsersController {
     }
 
     public void doDeleteUser() {
-        Alert deleteAlert = new Alert(Alert.AlertType.WARNING);
-        deleteAlert.setContentText(String.format("Weet je zeker dat je gebruiker %s wilt verwijderen?",
+        Alert deleteAlert = new Alert(Alert.AlertType.CONFIRMATION);
+        deleteAlert.setTitle("Verwijderen gebruiker");
+        deleteAlert.setHeaderText(String.format("Weet je zeker dat je gebruiker %s wilt verwijderen?",
                 userList.getSelectionModel().getSelectedItem().toString()));
+        deleteAlert.setContentText("Dit kun je niet ongedaan maken.");
+        Optional<ButtonType> result = deleteAlert.showAndWait();
+        if(result.get() == ButtonType.OK) {
+            userDAO.deleteUser(userList.getSelectionModel().getSelectedItem());
+        }
         deleteAlert.show();
-        userDAO.deleteUser(userList.getSelectionModel().getSelectedItem());
+
